requirements.txt
flask
requests
python-telegram-bot
pandas
tabulate
python-dotenv

